Add Worker Pool concurrency.go
hi, I optimized this code by adding a worker pool. Previously, you called a separate goroutine for each task. When there are a large number of tasks, this approach consumes too many machine resources. The best approach in this case is a worker pool, which limits the number of simultaneously working goroutines.